pyproject.toml
[tool.poetry]
name = "poetry2setup"
version = "0.1.0"
description = "Convert python-poetry to setup.py"
authors = ["abersheeran <me@abersheeran.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.scripts]
poetry2setup = 'poetry2setup:main'

[tool.poetry.dependencies]
python = "^3.7"

poetry-core = "^1.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
